Close web contents modal dialogs on content load start
Address the failure of some web contents modal dialogs to close when
interstitial WebUI is displayed by closing the dialogs on content load start.
Remove existing ad hoc support in dialogs for closing on page loading in favor
of a uniform approach.

I'm using load start as the trigger for dialog close as this seems to be the
point in time at which the user first perceives the page to be changing.

Notes on the print preview changes: with this change the dialog is closed when
load starts and the initiator tab gets removed via RemovePreviewDialog, so there
is no longer a need to handle NOTIFICATION_NAV_ENTRY_COMITTED for the initiator
tab.  The PrintViewManager::PrintPreviewDone() DCHECK is removed since now
PrintViewManager::RenderViewGone() may be invoked and print_preview_state_ reset
due to WebContents closure before the dialog is destroyed and PrintPreviewDone()
is invoked.

BUG=240575

Review URL: https://chromiumcodereview.appspot.com/17500003