sync: Associate bookmarks without moving them
The revision r206572 modified the change processing code so it no longer
had to touch every bookmark within a folder when only a handful of them
had changed positions.  This was expected to be a big performance
improvement in change processing for users with large accounts.

The change modified CreateOrUpdateBookmarkNode() so it would always
initialize or update nodes in the right-most position within their
parent folder.  This had an adverse effect on the other caller of this
function, the bookmark model associator.  With the new version, model
association now needlessly moved all nodes to the right before moving
them back to their proper position.

This change breaks up CreateOrUpdateModelAssociation and inlines it into
its two callers.  This allows us to implement the right behaviour in the
right context (specifically, not moving bookmarks needlessly during
model association).  It also introduces the function
UpdateBookmarkWithSyncData() so we can share some common code between
model association and change processing.

The new model association call still invokes Move() fairly often, but
the bookmark model should be smart enough to realize when this is a
no-op and ignore it.

BUG=256623

Review URL: https://chromiumcodereview.appspot.com/18953002