[Android] Don't use keyed tag for AutoFill tests.
The version of the View.setTag API that takes a key mandates
that the key must be an application specific resource ID. This
is fine for chromium builds but breaks webview at
runtime, as WebView isn't itself an application.

Work around this issue by attaching a single tag without a key.

Review URL: https://chromiumcodereview.appspot.com/18624003