(re-land) Flush each thread's message loop on destruction of TestBrowserThreadBundle.
Flush each thread's message loop on destruction of TestBrowserThreadBundle.

All threads, real or fake, are now flushed in the order they are destroyed.
Previously we flushed all fake threads once and then the real threads, which
caused some deletion tasks to be dropped.

Fixes memory leaks in AutocheckoutManagerTest tests.

This is the same as r210457 which got reverted due to a CHECK failure in
MediaStreamDispatcherHostTest.GenerateStream. Trying to address that issue in
this CL.

BUG=247525
R=ajwong@chromium.org, jcivelli@chromium.org, xians@chromium.org

Review URL: https://chromiumcodereview.appspot.com/18978007