Remove incorrect DHCECK from Invalidations
The DCHECK was intended to make sure that we've bothered to fetch an
access token by the time we call UpdateInvalidatorCredentials(). That
seemed like a reasonable assertion.  However, if we're calling
UpdateInvalidatorCredentials with the intention of clearing the token
(which can happen when we receive NOTIFICATION_TOKENS_CLEARED) the
DCHECK will fail.

BUG=259320

Review URL: https://chromiumcodereview.appspot.com/18565007