Extension activity log database refactoring (step 1)
This is the first of several patches to refactor the activity log
database, policy, and make changes to the database schema.

We move control over the database schema to the activity log policy,
since different policies might need different database layouts for
whatever summarization is implemented.

Schema setup is implemented by a call from the ActivityDatabase back
into the policy object.  Possible callbacks are made explicit in the
ActivityPolicyCallbacks interface; in the future there will be
additional callbacks related to writing actions to the database.  The
lifetime of the policy objects is tied to the ActivityDatabase object.

The table layout is not currently changed and code for creating the
tables is still in the Action classes; that will be changed in a future
patch.

BUG=255730

Review URL: https://chromiumcodereview.appspot.com/18660004