Pass LatencyInfo to GPU side on Android.
A LatencyInfo object is created for an input event.

To measure the frame buffer swapping latency, it traverses through
the following path:
1. browser passes the LatencyInfo object to render via
  content::RenderWidgetHostImpl::ForwardTouchEventWithLatencyInfo;
2. render attaches the LatencyInfo object to a LayerTreeImpl object,
  and later content::CompositorOutputSurface::SwapBuffers sends
  the LatencyInfo object back to the browser side;
3. browser gets the CompositorFrame object with the LatencyInfo object
  in content::RenderWidgetHostImpl::OnSwapCompositorFrame,
4. later browser sends SwapBuffers message to GPU via
  OutputSurfaceWithoutParent::SwapBuffer in compositor_impl_android.cc,
5. after swapping buffers, GPU sends FrameDrawn message to browser
  with the LatencyInfo object on the PassThroughImageTransportSurface
  object.

There are two missing steps in the path for the LatencyInfo object to
be passed to the GPU side:
a) in step 3, the LatencyInfo object needs to be set on the LayerTreeHost
   object so that it can be sent to the GPU side;
b) in step 4, the browser needs to explicitly send the LatencyInfo
   object to the GPU side.

The change addresses issue a) by setting the LatencyInfo object
on the LayerHostTree object in render_widget_host_view_android.cc.
Issue b) is addressed in compositor_impl_android.cc by sending
the LatencyInfo object to GPU side explicitly.

BUG=249201

Review URL: https://chromiumcodereview.appspot.com/19026002