Add interactive_ui_tests to be run as isolated.
This ensures no regression happens in the corresponding
interactive_ui_tests.isolate. Last time it was tried and reverted in r195803,
browser_tests had issues on XP/Vista slave. But it's been a long time and I
don't recall the exact issue so trying again with the expectation to have to
revert this CL so investigation can be done offline. Trying interactive_ui_tests
instead in case it works.

R=iannucci@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/18034030