Make scrollable viewport size no longer depend on clip layer.
In non-solid-color mode, the scrollable viewport size calculation was
ignoring device_viewport_size and using the clip layer size instead, in
order to take into account non-overlay-scrollbars.  This was weird and
was going to behave incorrectly with desktop pinch zoom.  This patch
changes the non-overlay-scrollbar size to be added to the max scroll
offset instead.

NOTRY=true
BUG=259141

Review URL: https://chromiumcodereview.appspot.com/18202006