cc: Make TestEnded() true immediately after EndTest()
The test change here fixes differences between when did commit and draw
frame happens for single and threaded compositor tests.  This was
causing a difference in the update counts for layers, as one version
would commit an extra time because the test wasn't done when it was
called.

This is just pre-refactoring of tests to make future patches for this
bug cleaner.

R=danakj@chromium.org
BUG=256381

Review URL: https://chromiumcodereview.appspot.com/18179011