Histogram detailed errors for IndexedDB reads.
We previously had great detailed logging for database opens
and writes but nothing for reads.

Docs recently revealed that they are frequently getting read
errors. While fixing these is less of a priority than errors
on open and write, we should start collecting the data now
so that we'll have it when it is a priority.

BUG=225051

Review URL: https://chromiumcodereview.appspot.com/18560011