PNaCl: Have plugin's LoadNaClModuleCommon start sel_ldr from the main thread.
Down stream ppb_nacl_private's LaunchSelLdr expects to
run on the main thread to be thread safe. PNaCl currently
violates that, using arbitrary helper threads to run
LaunchSelLdr.

This changes the plugin to bounce the request to the
main thread and block until complete before launching
sel_ldr. We only do the sel_ldr process creation on
the main thread, and the blocking load/validate of the
nexe is split off and done on the background thread.

TODO: Remove the EnableBackgroundSelLdrLaunch(), since we are no
longer doing the actual work off the main thread.

BUG=https://code.google.com/p/chromium/issues/detail?id=259010

Review URL: https://chromiumcodereview.appspot.com/19092004