Fix configuration_policy=1 Android build.
The chrome/browser/extension/api files are compiled in on Android, and when
configuration_policy==0 then the managed_value_store_cache.cc file is excluded.

But when it is included, then it depends on the storage_schema_manifest_handler.

This regressed in http://crrev.com/211543.

BUG=247435

Review URL: https://chromiumcodereview.appspot.com/19157002