Delete host controllers when they report an error.
HostController instances used not to be deleted nor removed from the
HostController map when an error happened.
This could lead to resource leaks or make the host_forwarder believe that a
port was already forwarded although the corresponding HostController instance
was in a dead state.

This CL moves the HostController management to a separate class,
HostControllersManager, that operates on its own thread. This is needed since
the main thread never yields which makes task posting impossible on the main
thread.
HostController instances are now created on this internal thread and
deleted/removed from the controllers map on this same thread as well when an
error happens.

BUG=242846

Review URL: https://chromiumcodereview.appspot.com/18099005