[Android WebView] Disable preemptive InitializeHwDraw
Currently we request a call RequestDrawGL when we are attached to the
a hardware view tree to enable hardware mode in case we get a hardware
draw. However, there are apps that attach a webview to a hardware tree
but never make hardware draw calls. In this case, a lot of resources are
wasted and it can cause rendering glitches in cases where we are still
not handling GL calls well.

BUG=internal b/9864710
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/19286009