cc: Don't DeleteResource while holding write lock.
When the context is lost and we can't generate a mailbox for a
resource, we delete that resource. However currently we are already
optimistically holding a write lock on the resource.

Instead, only take the write lock once we have generated the mailbox
so we don't hold it when we delete the resource.

Tests:
VideoResourceUpdaterTest.SoftwareFrame
VideoResourceUpdaterTest.LostContextForSoftwareFrame

R=piman@chromium.org
BUG=260361

Review URL: https://codereview.chromium.org/19272005