[Telemetry] Don't stop the WPR server when we hit a page with a file URL.
This makes replaying page sets with mixed http:// and file:// URLs work. The
reason is that the browser is configured to send all traffic to the WPR
server regardless of the port in the URL. So the WPR server forwards the
request on to the local HTTP server when it gets a request with a host of
a different port. For this to work, WPR must continue to run.

BUG=257209
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/19480003