Enable pop-up menus in OSX App Launcher for search results.
Adds a lazily created MenuController for search result items in the
AppsSearchResultsModelBridge. This is populated from the model when a
menu is requested for a context-menu event on the results NSTableView.
The menu is the same as the context menu shown on the items in the grid.

Also fixes a bug where the lightweight ItemObservers were not created
when setting the initial model. Only really a problem in tests, since
the search results model is otherwise empty when first populating the
view. However, the observers should be updated at the same time to be
consistent.

Screenshot at http://crbug.com/138633#c78

BUG=138633
TEST=Search for an installed app in the OSX App Launcher. Right-click
the result. A (functional) context menu should be shown matching the
menu shown in the grid. Also covered by added
app_list_unittests:AppsSearchResultsControllerTest.ContextMenus

Review URL: https://chromiumcodereview.appspot.com/18071011