Add reset button for each preference in gesture config UI
A reset button is added for each preference in the gestue config UI
(chrome://gesture). When no value is set for a preference (i.e. it is
set to default), the button is disabled. When a value is set, the button
is enabled and can be used to reset the preference to the default value.

BUG=251858

Review URL: https://chromiumcodereview.appspot.com/18023026