This change adds the following fields from the ThemeService to the ThemeBackgroundInfo: - text color - text color light - link color - header color - section border color - using default theme - alternate logo
These values are used when a theme is present in the local NTP and are exposed via the API to the selected search provider as well.

This change also inserts a custom style element if a theme is present instead of using the custom-theme class. 
The current logo has been replaced by a logo with transparency.

BUG=222489, 246530

Review URL: https://chromiumcodereview.appspot.com/16320004