Add WebRTC DEPS for Android builds.
This DEPS makes it possible to pull required additional
dependencies for Chromium that are needed to build
and execute WebRTC Android tests using Chromium's
Android test framework.

In addition to this deps solution, custom deps are needed
for src/third_party/webrtc in order to checkout the latest
code instead of the DEPS pinned revision of WebRTC stable
branch.

BUG=webrtc:2059
TEST=checked in the file in a local Subversion repo and tested
adding it as a separate solution, followed by a successful
gclient sync, with a .gclient configured for Android (see
https://code.google.com/p/chromium/wiki/AndroidBuildInstructions
for details).

R=hellner@google.com

Review URL: https://codereview.chromium.org/19235014