Expose the method used for the next URLRequest redirect.
This makes it possible for URLRequest client code to determine
the method that will be used to resolve the next step in the redirect
chain.
This is to fix android_webview not doing navigation intercepting
if the result of a POST resulted in a 302 redirect.

BUG=b/9773847
TEST=AndroidWebViewTest, net_unittests

Review URL: https://chromiumcodereview.appspot.com/18429010