Omnibox: Search Provider: Fix Minor Constraint Enforcement Bug
IsTopMatchHighRankSearchForURL() is supposed to be check if
the top match is not a URL.  It currently does this by asking
is it search-what-you-typed, search-other-engine, or search-suggest?
If so, then it must not be a URL.

This seems to have led to a bug because search-history has been omitted.

I haven't seen this bug in the wild; hence there's no bug link.  I 
spotted it when looking at the code.

Rather than fix it by adding "or is search-history" to the test, I changed
the test to be a negative.  After all navsuggests are the only things
search provider can return that are URLs.

BUG=

Review URL: https://chromiumcodereview.appspot.com/18889008