IndexedDB: Make perf test scripts use a seeded PRNG
Some of the tests select keys to read randomly. Random in a test is
not a virtue due to caching. Switch to a simple seedable PRNG for
repeatability. I didn't notice any overall perf difference or reduced
variability on my machine in limited testing.

BUG=188746

Review URL: https://chromiumcodereview.appspot.com/19279003