net: allow fallback down to TLS 1.0 in the event of a bad-record-MAC alert.
TLS 1.1 support has uncovered several examples of a new kind of broken server:
they negotiate TLS 1.0 correctly in the face of a 1.1 or 1.2 ClientHello, but
then fail with a bad-record-MAC alert when processing the client's Finished
message.

This bug is exhibited by at least two different types of SSL "accelerator"
device, which will probably take forever to be fixed. So, with a heavy heart,
this change adds yet another workaround.

BUG=260358
R=rsleevi@chromium.org

Review URL: https://codereview.chromium.org/19607008