Fix page scale factor affecting transforms higher on the tree.
Page scale factor was applied in an unusual manner, via a
ConcatTransform on every individual layer affected by it.  This is only
correct if the scaling happens to be commutative with all transforms
above it.  This patch fixes it to a more usual "Preconcat" at the
correct place in the tree traversal.

I also changed page scale factor to be applied as a sublayer_matrix of
the layer above the root scroll layer.  (This is currently the root
container layer and will be the "page scale layer" when that is
introduced.)  This allowed fixed-pos compensation code to be simplified
to not consider page scale.

New test coverage in LayerTreeHostCommonTest.TransformAboveRootLayer.

NOTRY=true
BUG=260356

Review URL: https://chromiumcodereview.appspot.com/19419002