Delete DeviceController instances before HostController ones get deleted.
While the Forwarder class on the device could get deleted when it detected an
error on a socket, the corresponding DeviceController could stick around. This
resulted in an excessive amount of file descriptors staying open that could
make the net_unittests fail.

This CL adds an explicit unmap command that gets sent by the host to the device
right before the host controller gets deleted. When the device receives this
message, it deletes the DeviceController instance corresponding to the port
being unmapped.

BUG=242846
R=bulach@chromium.org

Review URL: https://codereview.chromium.org/19519007