Allow delete of multiple bookmark nodes in Bookmark Manager Private API.
This change is to support the multi-level undo for bookmark feature. In
order for the undo of the deletion of multiple bookmark nodes the
deletions need to be grouped as one user action.

With this change the grouping of the deletions will be accomplished in
the API call in the future, when the bookmark undo service is fully
integrated.

BUG=126092

Review URL: https://chromiumcodereview.appspot.com/19615002