cc: Remove invalid ancient dchecks
The original purpose of this dcheck was to make sure that we removed
render surfaces that wouldn't draw.  However, this dcheck is out of date
(e.g. copy requests) and is far away from the code it's checking.  We
should just remove it because it's not adding anything.

R=danakj@chromium.org
BUG=none

Review URL: https://chromiumcodereview.appspot.com/19547007