ash: Use correct cursor when dragging out a tab over the omnibox
Dragging a tab out of the tabstrip by dragging it down over the omnibox can
result in an I-beam cursor during the move-drag of the new window. This
happens because RunMoveLoop() from the TabDragController releases mouse
capture for the source browser window, which resets the cursor pointer based
on the last mouse position. This will be an I-beam if the mouse was over the
omnibox.

Since we don't ever want an old/out-of-date cursor during a window drag,
sanitize it to an arrow pointer for mouse-based window moves.

BUG=242118
TEST=manual, slowly drag out a tab downward over the omnibox

Review URL: https://chromiumcodereview.appspot.com/19035002