Allow noise TouchBegin to be dropped
We currently don't detect & drop noise TouchBegin, but only for noise
TouchUpdate & TouchEnd. This could result in a noise TouchBegin is
translated to be a TouchPress at (0,0) and if not cancelled quickly
enough, it triggers long press effect at (0,0). This CL changes to
drop the noise TouchBegin.

BUG=237899
TEST=2F Swipe in horizontally from side bezel, and check that no
     long press effect appears at (0,0)

Review URL: https://chromiumcodereview.appspot.com/19649005