sync: Add pre-commit update avoidance mode + flag
This CL allows the syncer to avoid requesting updates from the server
prior to a commit if the nudge tracker determines that such a request is
not necessary.  In general, the nudge tracker will consider such a
request to be unnecessary unless the sync cycle was triggered by a local
refresh request or an invalidation, or invalidations are disabled.

This behavior is enabled only if explicitly requested by the client or
the server.  The client can enable this mode with the command-line flag
'--sync-enable-get-update-avoidance'.  The server can toggle it on or
off with the 'pre_commit_update_avoidance' experiment.  Unless at least
one of these flags are set, this CL will have no effect on syncer
behavior.

When this experimental mode is enabled, the SyncScheduler will use the
"short poll" interval rather than the "long poll".  This should help
clients keep in sync with each other even if the experiment breaks.

Also included in this CL are some extensions to the python test server
to enable manual testing of the pre_commit_update_avoidance experiment
node.

BUG=147685

Review URL: https://chromiumcodereview.appspot.com/19309002