NaCl IRT: Remove now-unneeded IrtInit() definition and call
It used to be that this Chrome-provided function was called from the
NaCl side, but reportedly only in the x86 build and not the ARM build.

It is now consistently defined on the NaCl side and called there, so
we can remove the Chrome side's duplicated definition and call.

Also remove the unnecessary NaClSrpcModuleFini() call: We never really
need to finalise modules like this when exiting.  NaClSrpcModuleFini()
might only have the effect of breaking get_random_bytes() if it's used
after the PPAPI event loop exits (though I don't think we ever really
expect the PPAPI event loop to exit).

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3186
TEST=browser_tests, nacl_integration

Review URL: https://codereview.chromium.org/19479005