Fix border and tiling issues when rendering aura progress bars.
There are 2 problems addressed by this CL.  The first is that the 
borders were not being drawn at all because when adjusting their
width for the scale factor required by the source -> dest mapping
it was less than 1 and being truncated to 0.

The second issue is that since we draw a separator in the progress
bar, the width of each "section" may not be evenly divisible by
the width of the progress bar, and the last section would be cut
off.  The fix here is to draw a subset of the sections with a width
of 1 pixel less so that it fits exactly into the progress bar.

BUG=176221

Review URL: https://chromiumcodereview.appspot.com/19682002