Handle the case the list items are removed but the selection index is not changed.
Originally, the change event of the file list selection is dispatched only when
the index of the selection is changed. But it should be also dispatched when the
list item is removed because if a list item is removed, the item referredby the
selection index is changed.

This CL added a helper function that can force the selection to dispatch the
change event and handle the problem by using the helper function.

Along the way, this CL also fixed a bug that the keyboard event dispatched at the
initialization stage causes error messages.

BUG=253887
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/18319003