cc: Limit number of raster tasks that can be scheduled.
Adds a limit of 256 to the number of raster tasks that can be
scheduled. 256 should be high enough to not cause unnecessary
scheduling but gives us some insurance that we're not spending
a huge amount of time scheduling one enormous set of tasks.

This also includes a
s/CheckForCompletedTileUploads/UpdateVisibleTiles/
cleanup.

BUG=262335

Review URL: https://chromiumcodereview.appspot.com/19792006