[NaCl SDK] Bring back -Wall and fix or disable all warnings.
Add -Wall to all SDK project, but not to the default
common.mk system (we don't want to force this on external
users).

gtest generates warning in gcc which can be suppressed with
with a gcc-only flag that clang doesn't understand, so I
added the ability to set GCC-only CFLAGS.  This seemed
better than forgoing the benefits of Wall on all
gtest-using projects.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/19802008