Files.app: Save the folder shortcuts to chrome.storage.
Folder shortcuts (former pinned folder) are saved to chrome.storage. They are restored when the Files.app is relanched.

The folders in Drive are saved to syncable storage and they will be synced among devices. The other folders are synced among instances in the same device, not among devices.

BUG=263274
TEST=manual
R=mtomasz@chromium.org

Review URL: https://codereview.chromium.org/19332002