[Telemetry] Respect blink PerformanceTests/Skipped again after r212642.
This fixes the blink_perf tests on the perf bots.

The problem was that the code was moved and page_set_arg was renamed to
path. That stompled on this code's usage of the var path. So just rename
it to avoid conflict.

BUG=237412
TBR=dtu@chromium.org
NOTRY=True

Review URL: https://codereview.chromium.org/19497005