ChildProcessSecurityPolicy: Add CopyIntoFileSystem permission.
This change adds a CopyInto permission into the ChildProcessSecurityPolicy. This permission is used by the Media Galleries API write support, which we hope to get into M30.

Currently, it's implemented by granting Create. Eventually, we want CopyInto to be a more restricted permission.

This requires divorcing the ChildProcessSecurityPolicy class from PlatformFile access mode flags, which is an in-process CL. See:
 - Bug: http://crbug.com/262142
 - CL: https://codereview.chromium.org/19599006/

BUG=NONE

Review URL: https://chromiumcodereview.appspot.com/19639003