cc: Remove some DCHECKs in CopyOutputRequest code.
These DCHECKs fire if we ever commit to the impl side twice without
performing a draw in between. There are various ways that this
could happen, and it doesn't really matter to the CopyOutputRequest
code anyways, so remove the DCHECKs and deal with the case gracefully
instead.

Test: LayerTreeHostTestAsyncTwoReadbacksWithoutDraw

BUG=263486
R=enne,piman

Review URL: https://chromiumcodereview.appspot.com/19668008