[NaCl SDK] Check that browser has support for NaCl/PNaCl/PPAPI.
Demonstrate a canonical way to detect if the browser
has support for NaCl/PNaCl in the SDK examples.

Do the check before creating the embed tag, when the page
loads. If the browser does not have support, do not create
the embed and just update the status message.

If NaCl is disabled in about:plugins, then this will show
an error.  If NaCl is not enabled in about:flags, then
this will not show an error and continue loading.  The
embed tag will then either (a) be created for CWS apps
or (b) show a puzzle piece for the open web.

For PNaCl, the page will show an error and not attempt
to create an embed tag if either NaCl is disabled
in about:plugins, or PNaCl is disabled in about:flags.

BUG=243040
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/19751003