[Telemetry] Add a profiler based loading measurement.
This measurement reports the hottest functions during page load. I plan to run
it over the top million sites.

It works by loading each page twice, the profile is captured for the second load
only. This is because cross-origin navigations cause the renderer to be swapped
out during profiling and to get more stable results.

This required a couple of core modifications:
1. csv_page_measurement_results should work when the page runner creates a
   dummy results object because discard_first_results is enabled.
2. discard_first_results should discard the first result of each page in
   the page set. Not just the first page.

BUG=None

Review URL: https://chromiumcodereview.appspot.com/19857003