Ensure we don't crash if user navigates back from NTP to Chrome sign-in page before it has fully loaded.
This CL aims to be simple, so it can be merged back to M29. It changes from DidStopLoading, which is later than needed point in time, to DidNavigateMainFrame. It is safe if we have committed a navigation to delete the previous navigation entry.

I've also added checks to ensure that we don't try to remove a pending navigation entry, as it leads to bugs, as evidenced here.

A follow up CL can completely remove the CurrentHistoryCleaner object, as all it does is store a single bit of state - the history entry index, and monitors for the same navigation events as OneClickSigninHelper.

BUG=257277

Review URL: https://chromiumcodereview.appspot.com/19699007