Fix issues showing GPU memory in task manager.
Don't put GPU process usage (the total over all processes) in parens
because people thought this meant it was negative. Rather, put an
asterisk next to it.

Also, if multiple tabs are sharing a process, only show the memory usage
for the group once.

BUG=

Review URL: https://chromiumcodereview.appspot.com/19798005