[sql] Cleanup open and close error histograms.
Cleanup open-time histograms to use the sparse framework so that
extended histograms can be recorded - also add entries to the xml
file.

Add a histogram to track close-time failures.  This should never fail,
but if it is it probably leaks memory and file handles, and possibly
locks.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/18978012