Remove CloseFile from FileSystem.
CloseFile has a potential issue, which may close a file different from opened
one. To avoid such a case, by this CL, OpenFile starts to return a closure,
which should be called on the file closing.
This is the same technique used in fileapi implementation.

BUG=259184
TEST=Ran unit_tests and tested manually.

Review URL: https://chromiumcodereview.appspot.com/19596003