History: Keep focus inside overlay when it is visible.
When the "Remove selected items" overlay was visible, pressing
tab cycle the focus through all the regular page elements as if
the overlay was not visible. This fixes the problem, ensuring that
when the overlay is visible, hitting tab cycles through the controls
in the overlay only.

BUG=259309
TEST=Open chrome://history. Click the checkbox of one of the history
items, then click 'Remove selected items'. When the dialog is visible,
hit tab and ensure that the focus cycles between the 'Remove' and
'Cancel' buttons in the dialog.

Review URL: https://chromiumcodereview.appspot.com/19943002