[Android] Fix leak of ContentViewCore via Accessibility ContentObserver
Neglecting to call ContentViewCore.Destroy (which is possibly the case
when it's embedded in Android WebView - we're at the mercy of the
embedding app) will cause the ContentObserver registered to watch for
changes to accessibliity script injection to leak, which includes
a native JNI strong ref back to the ContentViewCore. Oops.

Fix this by also clearing the ContentObserver in onDetachedFromWindow.

BUG=b/9912534

Review URL: https://chromiumcodereview.appspot.com/19383003