Simple Cache: Get mtime from the cache directory early
Currently the index is loaded asynchronously while the backend is already
initialized and is able to modify the cache directory. This may lead to index
file being considered stale without a real reason.

To overcome this problem we take the mtime of the directory early during backend
initialization and then compare it to the index freshness timestamp. The index
in memory will continue to be fresh since we update it with latest knowledge
from the backend.

This is the first step to add precision into index freshness, following steps
are:
* move the index file into a separate directory to make sure the rename()
  operation itself does not make the index stale
* Save directory mtime to the index file itself to avoid missing updates while
  the index is being written to the disk
* Read the last index modification time without loading the whole index

Some preparations were also made here for the next step: moving the index file
to a subdirectory. Mainly it is about moving some knowledge of cache directory
structure into the SimpleIndexFile and updating the tests accordingly. Hopefully
this won't make reviewing too complicated.

BUG=261618

Review URL: https://chromiumcodereview.appspot.com/19519010