Avoid handling commands in Files.app when a dialog is displayed.
Before, we were handling commands, such as Ctrl-F for searching, even if the dialog such as share dialog or delete confirmation was displayed. This patch resolves this issue by registering the commands on the dialog container, which instead of on entire document. As a result, these commands are not invoked by shortcuts while a popup dialog is displayed.

TEST=Follow the bug's repro steps.
BUG=263703
R=hirono@chromium.org

Review URL: https://codereview.chromium.org/19734015