PNaCl: Fix IRT wrapper to return proper error value
-1 is not a valid error return value for the IRT query function, which
is supposed to return 0 if an interface is not found.

This is just a nit, because this error is not likely to occur in
practice, but the presence of "return -1" might lead people to think
this is a valid error value for this interface.

BUG=none
TEST=none

Review URL: https://codereview.chromium.org/20438002