PNaCl: Simplify IRT wrapper a little by removing a global variable
There is no need for the temporary interface struct to be assembled in
a global variable, ppapi_hook.

This also removes a use of memcpy().  memcpy() is not available in
this context because this file isn't linked against libc.  It only
works because the compiler inlines the memcpy() call.

BUG=none
TEST=PNaCl tests in browser_tests

Review URL: https://codereview.chromium.org/20508002