GPU: Use TLS pointer for GLApi.
This is straight from:
https://codereview.chromium.org/16894002/

To support GL access on multiple threads, GL-API needs
to be in TLS. It hasn't mattered previously in practice
because our limited threaded GL calls were always forwarded
to the right native calls. But as we improve virtual contexts
and unset the API on ReleaseCurrent, this will no longer be
the case.

BUG=261264,263398

Review URL: https://chromiumcodereview.appspot.com/20014006