Close open secondary widgets on exit
We keep track of all top-level windows by adding the window id to a list every
time a window is created and removing it when it is destroyed. When the browser
shuts down, we close any unclosed windows to prevent the browser from crashing.

BUG=263501

Review URL: https://chromiumcodereview.appspot.com/20017007